tidy(struct): move provider to gateway
更换了项目结构，不再将provider与gateway分离，gateway将直接行驶provider的职责
