+++ pkg/provider/provider.go
@@ -7,18 +7,6 @@ import (
 	"github.com/miRemid/cqless/pkg/types"
 )
 
-var (
-	defaultProvider *Provider
-)
-
-func init() {
-	defaultProvider = new(Provider)
-}
-
-func Init(config *types.CQLessConfig) error {
-	return defaultProvider.Init(config)
-}
-
 type ProviderPluginInterface interface {
 	Init(*types.CQLessConfig) error // 初始化Plugin
 
@@ -30,15 +18,4 @@ type ProviderPluginInterface interface {
 	Resolve(ctx context.Context, functionName string) (url.URL, error)
 
 	Close()
-}
-
-type Provider struct {
-	plugin ProviderPluginInterface
-}
-
-func (p *Provider) Init(config *types.CQLessConfig) error {
-	if err := p.plugin.Init(config); err != nil {
-		return err
-	}
-	return nil
 }
